More inventory work, fix one packet, misc cleanup

+++ src/main/java/net/rush/packets/packet/HeldItemChangePacket.java
@@ -34,6 +34,6 @@ public class HeldItemChangePacket extends Packet {
 
 	@Override
 	public void read17(ByteBufInputStream input) throws IOException {
-		slotId = input.readByte();
+		slotId = input.readShort();
 	}
 }
